refactor(check-mark-white-icon): fix typo in component name
